rocksdb: Build with jemmaloc allocator
We use jemalloc for memory tracking, so we need to build rocksdb with
jemalloc as well so that memory used by rocksdb is properly accounted
for. This particularly matters for evictions--if we don't properly
account for all memory used by the process, then we could possibly fail
to evict and the process could be oomkilled.

Change-Id: Ied8ce13e046e9d5025e453db140945362ad631e3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5898
Tested-by: Buildkite CI
Reviewed-by: Aspen Smith <aspen@readyset.io>