replicators: MySqlPosition::position overflow fix
This is actually a flaw in MySQL binlog protocol, as Binlog::EventHeader
defines end_pos filed as 4 bytes, while binlog might grow beyond that
point. When that happens, the end_pos offset delta above u32::MAX will
be reported. If the said event is from a table, readyset is replicating
from and the last known position on that table is higher than the
reported overflown position, we will skip the event:

2023-08-27T17:48:35.713716Z WARN replicators::noria_adapter: Skipping
table action for earlier entry table=test.user
pos=binlog.000003:384346131 cur=binlog.000003:4289370691

This causes readyset copy of the table(s) to be out of sync with the
source database.

This patch fixes the issue by using u64 for the position field in
MySqlPosition.

MySQL does not have an API that allows us to register a replication
stream to start streaming binlog from the overflow position. We will
have to re-snapshot all tables if such situation happens.
The current patch does not address this issue. We will panic in case
unwrap() fails.

Refs: #465
GitHub-PR: #499
Change-Id: I0d247c6c83f20729be5458faccfeb06d30e4fd0f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5944
Reviewed-by: Aspen Smith <aspen@readyset.io>
Tested-by: Buildkite CI