jepsen: Switch to a simpler workload
Only a week into jepsen testing, and I've already hit up into the
internal inconsistency problem. This probably shouldn't surprise me.

Writing an efficient checker for eventual consistency of a query with
joins when we don't have internal inconsistency is *hard*. We might
still do it at some point, but for the time being I'm switching to a
simpler workload that still exercises a few properties of the system - a
single count(*) with a group and a parameter - and I'm going to start
expanding on nemeses to try to introduce *some* kind of breakage.

Change-Id: I2677e4f8e977df0d0e733b3b8f043de9d338b559
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5905
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>