jepsen: Nodes might die before the kill op finishes
When we decide to kill an adapter node, the process might start failing
to reply to queries *any time* between when the kill op starts and when
it finishes. The liveness checker was incorrectly failing in this case,
though -  if it found `:fail` ops *any time* before the kill op for the
last live adapter finished, it'd mark the history as invalid. Now,
instead, we track how many kill ops are currently pending, and allow ops
to fail if they happen when we are killing as many adapters as we
currently think are alive.

Fixes: REA-3402
Change-Id: I10fc74f43124b7a753325eef41469c408144b5a2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5907
Reviewed-by: Nick Marino <nick@readyset.io>
Tested-by: Buildkite CI