replicators: Use end LSN of commit as pos
The `lsn` field in COMMIT XLogData messages matches the LSN of the last
event in the transaction, meaning the event before a given COMMIT and
the COMMIT itself will share the same replication offset (specifically,
`(LSN of commit, LSN of commit)`). While it's okay for two non-COMMIT
events to share the same replication offset, if the event before a
COMMIT has the same replication offset as the COMMIT itself, and we
flush the events in our buffer before we process the COMMIT, we'll
ignore the `ReplicationAction::LogPosition` we return when we see the
COMMIT and log a "Skipping schema update for earlier entry" warning.

This commit uses the `end_lsn` field in the COMMIT XLogData message as
the second component of the COMMIT's replication offset to ensure that
the event before a COMMIT and the COMMIT itself share the same offset.

Change-Id: Icf5e6ad53956b81cbdb5345e3d65723764a7fa75