replicators: Don't add 1 to pos before ACK
Previously, we were adding 1 to the position we ACK in our standby
status updates to Postgres because the Postgres docs describe the
position in that status message as "the location of the last WAL
byte + 1 flushed to disk in the standby".

My understanding at the time of this commit is that the status update
message is documented this way to allow for replication receivers always
to report their current position and the position given in keepalive
requests without having to worry about whether they've seen every event
at that position. In other words, the Postgres WAL sender understands
that when a replication receiver ACKs a given LSN, it understands that
to mean that the receiver has seen every event *up until* that LSN. The
documentation describes the position as including a "+ 1" to account for
this "off-by-one" behavior. Even the Postgres logical replication receiver
itself does not add 1 to the positions it reports in status updates.

This commit removes the "+ 1" from our standby status updates to be
consistent with what Postgres does.

Change-Id: I13842e2d2bddf0e71894b5484caa5c688c95321e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5913
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>