replicators: ACK LSN from keepalive request
Previously, we always reported the LSN of the last event we applied to
base tables as our current position when responding to keepalive
requests. This meant that if the database we were connected to wasn't
receiving any writes but *another* database on the same server *was*
receiving writes, the WAL would continue to grow, but we'd continue to
ACK the same position. This led to replication lag, as Postgres thought
that it had to keep around all of the WAL entries from the writes to the
other database.

Keepalive requests from Postgres include an LSN that refers to the point
up to which events have been sent to us so far. Even if there are writes
occurring on another database on the server, we can use this LSN to
acknowledge that we've applied all of the relevant events up to that
point. This commit makes use of the LSN from the keepalive request by
reporting it as our current position as long as:

1. We are not in the middle of processing a transaction, which is
   important to ensure that we don't ACK a position that occurs outside
   of the transaction we are currently in; and
2. We have no actions buffered, which is important because if we had
   buffered actions, ACKed the position in a keepalive request, and then
   crashed, we would have ACKed a position past the point at which we've
   actually persisted events to base tables

Fixes: REA-3325
Change-Id: If8419e22c6647e9e6d34c7cba13829efd872dac0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5914
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>