replicators: Fix "until" condition
`PostgresWalConnector::next_action` takes an `until` parameter that
allows the method to return early if the we've reached the position
referenced by `until`. We were checking this condition after the point
at which we pulled the next event from the WAL stream, so if the
condition was triggered, we'd lose the event. To avoid this, this commit
moves the `until` check to be *before* we pull another position from the
WAL.

Change-Id: Ibbe6478844e2064c61256d88b277d460900eca8d