adapter: Better error when ReadySet user/pass missing from upstream url
Print a more descriptive error message when ReadySet tries and fails to
infer a username and password (for clients to use to connect to
ReadySet) by parsing the upstream database URL. It attempts to do so
when the --allow-unauthenticated-connections option isn't set, and
--username and --password aren't provided.

This is a bit of a corner case. It is possible for the user to do the
following:
- not require credentials to connect to their upstream database
- require credentials to connect to ReadySet

In practice, though, this makes little sense. In all likelihood, if a
user hits this condition, and opts to provide -u/-p, ReadySet will then
fail to connect to upstream because of missing credentials, which is
handled separately.

Change-Id: Ief7cd77421513b29713bea1e801f7bf4104bbb46