jepsen: Apply writes on invoke, not ok
If things happen fast enough (which they do some of the time in
practice!) a query can see the effect of a write after it starts, but
before Jepsen reports the op as having finished. This updates the
eventual-consistency checker to apply writes on `:invoke`, not `:ok`, to
allow for this. Since the shape of the insert map is slightly different,
this also has to update the `memory-db` to turn the map into the shape
we expect.

Change-Id: I9dd3ded1e4e96d02cd470c71e497a06ba232de55
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5917
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>