readyset-tracing: Allow disabling colored output
Sometimes we run in an environment where we want human-readable (eg
`full`, `pretty`, or `compact`) logs, but don't want ANSI escape codes
for colors (eg we're logging to a text file that'll get rendered
verbatim in a web browser, such as in Jepsen tests). This commit adds a
`--no-color` flag to `readyset_tracing::Options`, which allows disabling
colored output for all log formats for all binaries which use it (which
is basically all binaries right now).

Release-Note-Core: Add a `--no-color` flag to the `readyset` and
  `readyset-server` binaries, which allows disabling colored output in
  logs
Change-Id: If6a71fee98a8c6bedfd52c1bb08f32716f751610