jepsen: Make `:grouped-count` writes idempotent
Since we retry writes that report as failed, we can end up
double-applying the same write op if it fails after it's reported as
applied successfully by the upstream. This *might* want to be something
we address in the proxy itself, but for now it's simple enough to just
make writes idempotent by making `id` an actual primary key and running
with `ON CONFLICT DO NOTHING`.

This was incorrectly reported as a double-applied write within ReadySet
itself at #468, but was actually just us applying the same
non-idempotent write twice within the jepsen test.

Change-Id: Ib856cb4a6df65e1bae43e711783ccd7a18a8dc14
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5922
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>