jepsen: Retry queries twice, with a slight sleep
With the way we're killing the adapter currently, it's just too likely
that both retries of a query happen to get routed to an adapter that is
actively being killed. Bumping up to retrying twice (in addition to the
first try) with a slight sleep in between seems to avoid that case.

Change-Id: Ice77728d0f89c7cf7b1965ac93c1aea12af0e2a3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5925
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>