jepsen: Add a nemesis to bitflip rocksdb databases
Add a new nemesis to the jepsen test which picks a random file in a
random rocksdb database on the server node and flips bits in it at
random. This *ought* to trigger rocksdb's checksumming eventually, at
which point the table should fail and either resnapshot, or all queries
should go upstream.

Refs: REA-3369
Refs: #355
Change-Id: Idc35897fe92a346d8b6406a97c12c6c2962187fa