replicators: Added proptest suite for wal_reader.rs
Right now, we don't have any unit tests for any of the code in
wal_reader.rs.
As part of the efforts to fix REA-3170, this commit adds a proptest in
order to make sure that `into_noria_vec` returns the correct error.

This currently only tests with random `DfType`s (and a corresponding
`DfValue` for each of them), so it doesn't cover other types of errors
that might occur, like parsing ones.

This commit also makes slight changes to the arbitrary code we had to
generate `DfType`s and `DfValue`s (based on the type).
Some of that fine tuning was done in order to generate values that made
more sense for tests (like limiting the recursive depth of `DfType` to
2, so that we can have array of arrays, but not further nesting), and
some to fix some bugs (like using `proptest::sample::select` on empty
vectors, which would panic).

Change-Id: Id6de3e1bb2bd3509f69797220d8fa55ba29f099b