readyset-data: Add format field to PassThrough struct
This commit adds the new field, but doesn't do anything interesting with
it yet. The plan is to use this to support passthrough of text-formatted
parameters of unsupported types when we proxy queries, but for now we
just hardcode the format to Binary everywhere so as to maintain the
existing behavior.

Followup commits will add support for encoding parameters as text in
proxied queries when the format flag is set to do so, and once that's
done we can actually set the format flag to Text and use the PassThrough
type in the `get_text_value` function in the psql-srv decoder.

Refs: #266
Refs: REA-3183
Change-Id: I3419ba96e1435b30b08f2bd154832da19afd2d59