readyset-data: Implement encode_format for DfValue ToSql impl
The default implementation of this trait method is to use binary format
for everything, and we still do that for almost all cases, but the new
passthrough text format is the one exception.

By implementing this, tokio-postgres will automatically send the right
flags in the Execute message to signify that we're sending text encoding
for any passthrough parameter that has the `format` parameter set to
Text. Since the ToSql impl already just encodes whatever raw data is
held in a DfValue::PassThrough, no further changes are needed to support
passing through text parameters.

Refs: #266
Refs: REA-3183
Change-Id: I04f6fa1e91a153669c72f76cd44e98298a102533