replicators: Fixed other potential bugs related to WAL errors.
When calling `TupleData::into_noria_vec` (in `wal_reader.rs`), we have
to make sure that when we error, we return `WalError::TableError` for
all _foreseeable_ errors (parsing, conversion, etc), so the replicators
can mark the table as unsupported.
Failing to do so would make the server get into a restarting loop (read
from WAL, process data, error out, restart, get the same record from
WAL, and so on).

This commit checks that we are correctly mapping the errors in some of
the operations we perform, that return a `Result`, to make sure we are
compliant with that.

Change-Id: Ie496b935a4282161fbe7070459131ac8dd56f13b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5929
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>