adapter: Display metrics in SHOW statements
Displays query latency quantiles in SHOW PROXIED QUERIES and SHOW CACHES
if `--prometheus-metrics` is enabled.

Introduces a `MetricsHandle` in the adapter, which wraps a
`PrometheusHandle`, providing some ReadySet specific convenience
functions.

The `MetricsHandle` can be used to query quantiles for a particular
query id. The quantiles are returned from the snapshot of histograms
that was taken at the last call to `snapshot_histograms`. This is done
because the `PrometheusHandle` only supports returning all histograms
for a particular tag.

Release-Note-Core: ReadySet will display latency quantiles in SHOW
  PROXIED QUERIES and SHOW CACHES if `--prometheus-metrics` is enabled.
Change-Id: I30d50aff79a087ad497c26db8b9a03d285b6a690