server: Bail if soft file descriptor limit can't be set to desired value
Attempt to raise the soft file descriptor limit by doing the equivalent
of `ulimit -Sn 32768`, and bail if we can't i.e. if the hard limit
`ulimit -Hn` is smaller. If the soft limit already exceeded the desired
value, do nothing.

On Windows, this is expected to silently succeed, as it doesn't have an
equivalent resource limit. There may be other applicable limits which if
hit, would not be caught by this commit, and would cause ReadySet to
crash during compaction. Note that ReadySet does not currently support
any Windows platforms.

On macOS, `ulimit -Hn` may not yield the correct value for the hard file
descriptor limit; `sysctl kern.maxfilesperproc` should be used instead.
This commit handles this case correctly.

Change-Id: I481bf8251e5dd773d02dea206b36b66e478f28f1
Refs: REA-2899
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5942
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>