build: Move public build down into public dir
To support buildkite CI builds out of the github repo, everything
necessary for the CI run needs to live in the github repo.  This update
moves the pipeline and all supporting scripts and files down under the
public/ tree.  The main top-level pipeline is adjusted to run the
public/.buildkite/pipeline.yml on changes to public/.

NOTE: This copies a lot of files from outside public into public, which
means there's a bunch of duplication (build-image.sh, run-tests.sh,
etc.).  To remove that duplication, all references from outside public
would need to be changed to point to the copies in public, and the
non-public files deleted.

Change-Id: Ibaaf03f3e5189b275bfe21b061da78dc84e3b983
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5950
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>