replicator: Handle PG WAL rotation
If the number of WAL files on the upstream Postgres instance exceeds
`max_slot_wal_keep_size`, the database will mark them for deletion,
giving our replication slot a status of "unreserved." When the next
checkpoint happens, these files will be deleted, giving our slot a
status of "lost." This would typically happen if we aren't replicating
changes as quickly as they are being written upstream; the LSN we report
upstream would grow further and further from the current LSN on the
server, forcing it to keep a longer and longer history of WAL.

Previously, when this happened, we'd enter into an infinite retry loop
in the replicator: replication would error out and upon restarting the
replicator, we'd fail to start replication because our slot was
invalidated. This commit fixes this by querying the WAL status of our
slot via `pg_replication_slots` before we invoke `START_REPLICATION`. If
the status of our slot is "lost," a full resnapshot is triggered.

Fixes: REA-3303
Change-Id: I05b7ad23d92ac8f08b9b2d2cf83703f4c0583445