Revert "mir: Don't make join keys during decorrelate"
It might be perfectly fine to push filter nodes below INNER JOINs, but
in the case of LEFT JOINs doing so actually changes the
semantics (because we won't get rows with NULLs for columns on the right
anymore). For decorrelating these queries, then, since we need to be
turning join-key-like filters into join keys on LEFT JOINs, it's simpler
to do that unconditionally for *all* joins, rather than having to have
different behavior for INNER JOINs.

This reverts commit a5df0f377c82fbe6a6be4dff96c0ae0a912bc7aa, with some
changes to TODO comments since this likely won't be changing back

Refs: REA-3417
Refs: #461
Change-Id: Ic616ed7a89078eb78ad867a409e90f3f80e96dc9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5967
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>