jepsen: Fix test crash on first-time startup during upstream teardown
The first time we try to run the teardown step on an upstream DB node,
we will fail because Postgres hasn't been set up yet, and so we end up
hitting an "unknown user postgres" error which is not caught in the
catches that currently exist.

This prevented me from making it through the README instructions when I
tried to follow them last week, and adding this line fixed the problem.

Change-Id: I1ef75d25f2fcae5af27bc9dbbe073bd1efeb7cdf
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5972
Reviewed-by: Aspen Smith <aspen@readyset.io>
Tested-by: Buildkite CI