dataflow-state: Update insert methods to return duplicate row count
Instead of just returning `true`/`false` based on whether the row was
inserted, this changes some insert-related code to now return `Some(n)`
or `None`, where `Some(n)` means the row was inserted and gives the
number of duplicates of that row that now exist in the state.

This doesn't make any user-visible changes, since the only code that
used the return values to these functions is also updated by this commit
to just call `.is_some()` on the return value. But, this lays the
groundwork for followup commits that will fix some bugs relating to weak
indexes.

Refs: REA-3336
Refs: #286
Change-Id: I71d146c893728099e0be85afbfdbb9eb3a40e5b6