benchmarks: Allow sending queries directly to upstream
Add ability to workload_emulator to send bench query traffic
directly to the upstream database. This allows easier
benchmarking across multiple scenarios (including using a
look aside cache, coming later in this CL).

Additionally, this benchmark now accepts a DataGenerator param,
which allows it to set up tables and data in the setup() function.
The prior behavior of the setup() function, migrate/prepare queries,
is moved to the benchmark() function. This change allow us to use the
--skip-setup parameter to avoid creating tables/data on each run,
yet still have queries migrated properly (if need be).

Change-Id: I3056af458fad31b04f3c8cfec8f21f48fc9d04b3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5994
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>