benchmarks: Add look-aside cache support (redis)
Expand the workload_emulator to be able to use a
look-aside cache similar to real-world uses: on reads,
attempt to find the key in the cache, and if found
return early. If the key is not found, read from the
upstream database, set into the cache, and return.

This commits adds support for redis; memcached will be added
in another commit.

There is a configurable TTL value, as a way to simulate
entries falling out of the cache. When setting the TTL,
though, we add a bit of jitter so that everything doesn't
expire from the cache all at the same time.

Change-Id: I3b19c6916d404d133876ba876a62251421f36105
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5995
Tested-by: Buildkite CI
Reviewed-by: Ethan Donowitz <ethan@readyset.io>