build: Fix build-image step dependencies
The `build-image` step is a dependency of both the public build pipeline
as well as `deploy-docs`, which only runs on the monorepo.  So we need
to run build-image in the public pipeline only if it hasn't already been
run for the monorepo pipeline.  This approach pulls the public build
itself into a `public-common` pipeline.  The monorepo pipeline can
define a `build-image` step and include `public-common`.  Likewise, the
`public/pipeline.yml` can define its own instance of `build-image` and
then include `public-common`.

Change-Id: Iedb470cb0de74722075a05b05e03e42116bd065f