benchmarks: update reqwest version due to winreg conflict
Updating, or forcing a base version, of reqwest, at least as far
as the benchmarks crate is concerned. This is because cargo deny check
was failing with multiple versions of winreg (which is a windows
registry reading lib). winreg is a transitive dependency of several
of our deps, the most obvious being reqwest. Updating to 0.11.19+
allows us to resolve the dependency conflict.

This change was necessary to faciliate using a memcached client
library in workload_emulator, introduced in another commit in this CL.

Change-Id: I2206b1687c8a2ac7260c709d2a39ed6d2b4efef3