adapter: Mark queries unsupported after max dry run time elapses
In explicit migration mode, mark pending queries as unsupported if their
dry-run fails for longer than MAX_PROCESSING_MINUTES (default 15).

It was already impossible for pending queries to become supported beyond
that duration (as we give up on dry-running them), but we would leave
them pending indefinitely. This allowed the user to try to cache them
explicitly; a failure here would indeed mark the query unsupported.

Refs: REA-3383
Change-Id: If79c19640edec0a60e29374125ffce1d50a6a6aa
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6007
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>