controller: Run recovery 20 seconds after workers die
When a worker dies and comes back, there are two possible orderings of
events:

Ordering A:
    1. The worker dies
    2. The worker restarts
    3. The controller is notified of the original worker's death

Ordering B:
    1. The worker dies
    3. The controller is notified of the worker's death
    2. The worker restarts

Currently, to avoid unbalanced domains in ordering A, we don't run
recovery on worker death - instead, we wait to run recovery until the
new worker registers itself with the controller. But in ordering B, when
the worker registers itself with the controller, the controller still
thinks the original (actually dead) worker is still healthy and running
all its domains, so doesn't schedule any of those domains onto the new
worker. There's a lot of ways we could fix this (see [0] for some
discussion) but the simplest tradeoff seems to be just running recovery
on worker death, after sleeping for an interval longer than twice the
time it takes to be notified of a worker dying. This achieves the
desired domain balancing in ordering B assuming the events happen
quickly enough (which they ought to) but also avoids the case where we
permanently have some unscheduled domains in ordering A.

To make sure we don't run more than one of these background recovery
tasks at once if more than one worker dies at the same time, there's
also a `background_recovery_running` atomic bool which we check before
running the recovery.

Fixes: REA-3020
Fixes: #217
Change-Id: Ib26d4b7bd47ee406c1fd43a24e7871f91bd43cc4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6008
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>