sql: Remove from view_schemas when removing views
Since we use the contents of `SqlIncorporator::view_schemas` to desugar
queries referencing views (eg to populate column lists in
`star_expansion` for queries which run `SELECT * FROM v`), it's
important that when SQL VIEWs are removed we also remove from that map -
otherwise queries referencing another view with the same name but
different column names would get the column names in the other view for
its `*`.

This commit does that removal in two places - first, when explicitly
dropping views eg as a result of a replicated `DROP VIEW` statement, and
second when views are dropped because they referenced tables that were
dropped. Note that in the second case we don't actually care about
`CASCADE` - the semantics of ReadySet have *all* drops imply `CASCADE`
and we just trust the upstream db to have already checked referential
integrity for us.

Fixes: REA-3284
Change-Id: Iec4569f4479221e016ffc9e656230cb6b08f2aef
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6012
Reviewed-by: Luke Osborne <luke@readyset.io>
Tested-by: Buildkite CI