nom-sql: Add support for parsing interval type.
Added support for parsing the interval type to nom_sql/src/sql_types.rs. Currently,
this type is unsupported by ReadySet dataflow, but this change prevents an internal
server error when the interval is parsed as an INT. Since the interval is a complex type,
a specific function was added to check for it before the integer types
are checked.

Change-Id: I4d629b6e2b555d9a9cc2083e1699b41384895c19