adapter: Introduce LazyUpstream
Introduce a new LazyUpstream type to upstream_database, which wraps
another type that impls UpstreamDatabase but makes it wait to actually
connect to the upstream until an upstream query is actually run.

The idea here is to save on upstream connections for client connections
that are only going to run a ReadySet-specific command by avoiding
connecting to the upstream DB until we know we're going to need to.

Refs: REA-3432
Refs: #496
Change-Id: Ibbf2e84ac93c23eaa79203b8554c359ecda64dbc
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6023
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>