adapter: Introduce LazyUpstream
Introduce a new LazyUpstream type to upstream_database, which wraps
another type that impls UpstreamDatabase but makes it wait to actually
connect to the upstream until an upstream query is actually run.

The idea here is to save on upstream connections for client connections
that are only going to run a ReadySet-specific command by avoiding
connecting to the upstream DB until we know we're going to need to.

Change-Id: Ibbf2e84ac93c23eaa79203b8554c359ecda64dbc