adapter: Only load schema_search_path once
Currently, we always load the schema_search_path in the upstream upon
initial client connection. Doing this prevents us from effectively
waiting to connect to the upstream until we absolutely *have* to in
order to run a query, since we'd end up just connecting to the upstream
immediately upon every connection anyway. This commit trades off the
ability to respond to changes in the server-global schema search
path (which isn't *that* necessary or common, at the end of the day) for
the (eventual) ability to avoid an upstream connection on the majority
of connections, by loading the schema search path *once* on adapter
startup, and reusing that for all connections.

The actual way we load the schema search path here is a little tricky -
since we don't want the adapter to crash if it can't connect to the
upstream on startup, we have a function which attempts to load the
schema search path once, and if it can't it retries in a loop in the
background until it can, storing the result in an RwLock.

Refs: REA-3432
Refs: #496
Change-Id: Ie7718f163634d91d804c93a91644bde10610964f