adapter: Use LazyUpstream for both PG and MySQL
Wrap the upstream DB for both the PostgreSQL and MySQL adapters in
LazyUpstream, so that they wait to establish a connection with the
upstream DB until they know they're going to need to proxy a query
upstream. This is generally a decent seeming idea, but is going to
become especially important in hosted environments that query things
like `SHOW PROXIED QUERIES` or `SHOW CACHES` frequently, to avoid that
placing additional load on the upstream database.

Fixes: REA-3432
Fixes: #496
Change-Id: I9509e2c8a67378651ecad5f178905886d01a633c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6027
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>