build: Move cargo-deny image creation to fix dependency graph
The cargo-deny image was being created if `public` was changed, but not
if `crates` was changed without `public`.  However, the `crates`
pipeline requires the cargo-deny image.  This update adds the cargo-deny
image build to the monorepo pipeline so that it will be in scope for the
`crates` pipeline even if `public` wasn't changed.

Change-Id: I6e0e22ed92ea590ef71685f94529f9f750a3bae0