dataflow: Add min persisted offset RPC
This commit adds a `/min_persisted_replication_offset` controller RPC
that queries each base table domain for the replication offset up to
which the table has persisted data to disk, computes the minimum of
those, and returns the result. Eventually, this offset will be used by
the replicator to determine the greatest offset it can ACK upstream such
that the upstream database doesn't purge any WAL files it would need to
replay events after a crash.

Refs: REA-3434
Change-Id: Id9e5f5dc0a2becd7e5b4f36a9041252dc5be8399