replicators: ACK min `persisted_up_to`
This commit has the Postgres replicator query the controller for the
minimum persisted replication offset whenever it needs to send a status
update upstream. If this value is present, it ACKs it upstream to tell
the upstream database that we've persisted all the data up to that
point; otherwise, it ACKs its current position.

Refs: REA-3434
Change-Id: I8c21f005af90a72b67d550a5b566452e636ffebf