persistent: Move db to separate `RwLock`
In a future commit, we'll need to be able to invoke `DB::sync_wal` in
such a way that doesn't block other writes. Although RocksDB writes
don't require a write lock on the RocksDB handle, writes through
persistent state that set a replication offset *do* require a write lock
on `SharedState` to set the replication offset. Because the RocksDB
handle and the replication offset live under the same `RwLock` in
`SharedState`, it is currently not possible to acquire a lock on the
RocksDB handle without a lock on the replication offset.

To rectify this, this commit moves the RocksDB handle out of `SharedState`
and into its own field on `PerisistentStateHandle` surrounded by its own
`RwLock`. This makes it possible to hold a read lock on the RocksDB handle
without holding a read lock on `SharedState` which would allow another
thread to concurrently acquire a write lock on `SharedState`.

Refs: REA-3434
Change-Id: I4b1f9f08af729b9d9e34c56593cc1f4c427b052f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6070
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>