nom-sql: Parse integer literals as signed by default
If we are parsing a raw integer literal and it doesn't have a sign, we
previously defaulted to interpreting it as an UnsignedInteger. This is
different than the behavior of mysql, which defaults to signed integers.
Postgres doesn't have unsigned integer types, so it makes sense for the
default to be signed integers.

Change-Id: I4bf9397255a305b50fd399afd2e95bf59d0e3d68