nom-sql: Add extra cycle to format parse round trips
Some literals can be valid for multiple types, so if we are just parsing
a query and don't have any types to associate with a literal, we will
default to one. This makes a simple format-parse round trip fail if
another invalid type is our starting state.

To avoid this, add an extra format->parse round trip and assert that the
2nd and 3rd parsed structs are equal rather than our initial state. This
lets us transform our initial arbitrary state into the initial state
with any ambiguously displayed types as their default types, even if
that default changes in the future due to parsing modifications.

Change-Id: Ia4f9d33b2e11f9abebc05cd24225aee7e0f7b07f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6052
Tested-by: Buildkite CI
Reviewed-by: Aspen Smith <aspen@readyset.io>