mir: Implement a predicate pushup pass
Add a MIR rewrite pash to push filter nodes as high up the graph as
possible (eg as long as they commute with the node they're being pushed
above and all columns in the filters resolve). The idea here is that
we'll hopefully push these filters above *joins*, meaning they can
restrict the list of rows that get sent to those joins, keeping the
internal state of the join smaller.

We call this "predicate push*up*" because we conventionally draw graphs
with bases at the top, but traditional database literature usually calls
this "predicate pushdown", since traditionally query plan graphs are
drawn the other way around.

Change-Id: I9d588239db86fdcaae6a490750ab0c98b1e370ed