readyset-client: Implement removal of stored caches
This adds a function to Authority to allow for removal of a stored
`CREATE CACHE` statement, for use when a cache is explicitly dropped,
along with a function for removing them all if `DROP ALL CACHES` is run.

Since removal of a cache uses a name of the cache (there can be multiple
due to cache aliases), we first resolve the alias to the original name
of the cache and use that to remove it from the authority. This matches
how we store the caches in the authority as well, and there is a
debug_assert added to enforce this invariant.

We specifically don't rely on the type system to enforce it because
keeping things as a String makes this storage less prone to backwards
incompatibility.

Change-Id: Id23eb198fd2aa8ebdfd83e7885aa82864b520adb