state: Propagate WAL flush errors upwards
This commit adds the ability for the WAL flush background thread for
persistent state to report errors upwards to the replicator by setting a
new `last_wal_flush_error` field on `SharedState`. This field is read
when the replicator queries for the "min persisted up to" offset, and if
an error is present, it is returned by the controller RPC. Upon
receiving this error, the replicator enters its retry loop, sleeping for
a brief period before reconnecting to the upstream database and
re-initializing replication.

Note that there is no change in behavior here from the previous state of
the world. Before we introduced background WAL flushing, errors that
occurred during the flushing/syncing of the WAL happened *in the write
path*, which means that the table RPCs sent by the replicator would
fail. This resulted in the same retry logic as is introduced in this
commit. In the future, way may want to do something more sophisticated,
such as parsing the error description we get from RocksDB to try to
understand why the flush/sync failed.

Refs: REA-3434
Change-Id: I069fcd2f654dace28f064d2240939d674a27532c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6072
Reviewed-by: Luke Osborne <luke@readyset.io>
Tested-by: Buildkite CI