replicators: Configure status update intrvl
Adds a configuration option that lets users configure the interval on
which we send updates to the upstream Postgres instance. This is notable
because we include in the status update the min LSN up to which we've
persisted data to disk across every base table. For higher write
throughputs, users may want to configure a lower status update interval to
ensure the size of the replication slot remains small.

Refs: REA-3434
Release-Note-Core: Added a config option to allow users to configure the
  interval upon which we report our current position to Postgres
Change-Id: I285cb1c658324178edadfc10b5b79b6c6dfb826b