persistent-state: Periodically persist WAL
Previously, we both flushed and synced the RocksDB WAL for a given base
table with every write. This was very costly from a performance
perspective: a write to a base table involved an fwrite syscall, an
fsync syscall, and waiting for the bytes to be persisted to disk. While
this was happening, the replicator was completely blocked from processing
WAL events, introducing a bottleneck into our write path.

The commit introduces periodic, background WAL persistence via a
background thread in `PersistentState`. On some preconfigured interval,
the background task will flush the WAL (which blocks other writes to
RocksDB), sync the WAL to disk (which doesn't block other writes), and
update the relevant replication offsets so the replicator knows up to
what point data has been persisted for a particular base table.

Fixes: REA-3434
Release-Note-Core: Improved write performance during replication by
  writing data to disk periodically in the background
Change-Id: I0ee06dc1837588701a02a83a2991696cfa0e1635