quickstart: Use port 6034 for quickstart metrics
The quickstart runs ReadySet in standalone mode, so 6034 (the adapter)
endpoint is the default.

This also changes the quickstart compose files to map port 6034 so it is
available to run http requests against with localhost by default.

It updates the grafana dashboards to look for metrics on port 6034
instead of 6033, which wasn't working in standalone mode and was the
impetus for this change.

Change-Id: I9e5af4501437d8a3f5b0153610ec83e01447a09d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/6092
Reviewed-by: Ron Hough <ron@readyset.io>
Tested-by: Buildkite CI