replicators: Change WAL pos during snapshot
When we snapshot a Postgres database, we use the replication slot's
"consistent point" as our current position in the logs; however, when we
first start up we transform this LSN to the `PostgresPosition`
`(consistent_point, 0)`. This means that if we don't receive any events
before the next time we need to send a status update to Postgres, we
would report our current position as "0/0", since that's the LSN
component of the `PostgresPosition`.

This commit fixes this by transforming the slot's consistent point to
the `PostgresPosition` `(consistent_point, consistent_point)`, which
ensure that we we'll report our position as the consistent point.

Refs: REA-3303
Change-Id: Icbfab89ff4641bda850eeeb0500afc2da342080d