server: Add an RPC for loading info about materializations
Add a new controller RPC for loading info about all the materializations
in the graph, including the index of the materialized node, the node's
name and description, the indexes in the materialization, and its size.

This RPC will be (but is not currently) used for retrieving the info
displayed in the result of the new EXPLAIN MATERIALIZATIONS command.

Refs: REA-3517
Change-Id: Ib9bc87011db55b3e1b3e672d63608b3163566f9a